hardware_test/requirements.txt
# Hardware Test Application Requirements
# For Raspberry Pi GPIO and I2C communication

# Core GPIO library for Raspberry Pi
RPi.GPIO>=0.7.1

# I2C communication library
smbus>=1.1

# Optional: For enhanced I2C functionality
# smbus2>=0.4.0

# Development and testing dependencies (optional)
# pytest>=7.0.0
# pytest-mock>=3.10.0

# Note: These packages are typically pre-installed on Raspberry Pi OS
# If not available, install with:
# sudo apt-get update
# sudo apt-get install python3-rpi.gpio python3-smbus
